Added napsterX project 2000 lines of styles for testing.
Brought down parsing time from 500 seconds to 0.18 seconds.
- The issue is StringQueue's Character to String conversion. Apparently it is very heavy. Bye bye StringQueue.
- We are going to use array fast iteration to go over each characters.
- Buffer is now a String.CharacterView class which supports dynamic addition and memory reservation.

There may be more things we can improve wrt parsing, such as we are double-parsing when it comes to Function/BlockHeader.
The idea's Parser should pass a struct of known interesting characters such as dot (to indicate class), colon (to indicate pseudoClass) to this Function/BlockHeader, since it already has to keep track of it.